Implemented sweptAABB, not fully in use yet
Note: This breaks game in current state. EntityContainer now spatial
partitions with regards to velocity aswell but velocities aren't
multiplied with delta time so it freaks out regarding the high speeds.
Need to change the stored velocities to be treated by deltatime first.
Should be a small fix.

Collision tests and all other tests pass fine at the moment. One of them
will start failing when sweptAABB is back in again. The comment by the
test also make this clear